python/requirements.txt
mediapipe==0.10.9
opencv-python==4.8.1.78
numpy==1.24.3
pyserial==3.5